fix: reduce pkg/mcp/factorysession file count for CI pkg-file-count gate
