+++ pkg/mcp/factorysession/read_session.go
@@ -8,11 +8,35 @@ import (
 	"github.com/portpowered/infinite-you/pkg/factorysessionexecution"
 )
 
+// GetSessionInput is the MCP request shape for you.factory_session.get.
+type GetSessionInput struct {
+	SessionID string `json:"sessionId"`
+}
+
+// GetSession returns one durable Factory Session inspection read model through
+// the you.factory_session.get MCP tool.
+func GetSession(service factorysessionexecution.Service, input GetSessionInput) ToolResponse[factoryapi.FactorySessionDurableReadModel] {
+	if service == nil {
+		envelope := unavailableServiceErrorEnvelope()
+		return ToolResponse[factoryapi.FactorySessionDurableReadModel]{Error: &envelope}
+	}
+
+	sessionID := input.SessionID
+	result, err := service.GetSession(context.Background(), sessionID)
+	if err != nil {
+		envelope := readErrorEnvelope(sessionID, err)
+		return ToolResponse[factoryapi.FactorySessionDurableReadModel]{Error: &envelope}
+	}
+
+	mapped := apifactorysession.SessionReadResponseToAPI(result)
+	return ToolResponse[factoryapi.FactorySessionDurableReadModel]{Result: &mapped}
+}
+
 // GetResultInput is the MCP request shape for you.factory_session.get_result.
 type GetResultInput struct {
-	SessionID        string                              `json:"sessionId"`
+	SessionID        string                               `json:"sessionId"`
 	Mode             *factoryapi.FactorySessionResultMode `json:"mode,omitempty"`
-	IncludeArtifacts *bool                               `json:"includeArtifacts,omitempty"`
+	IncludeArtifacts *bool                                `json:"includeArtifacts,omitempty"`
 }
 
 // GetResult retrieves one durable Factory Session result through the
